booking-service/requirements.txt
fastapi
uvicorn
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
